+++ src/test/java/io/hhplus/tdd/Integration/currency/PointServiceFOTest.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.Integration;
+package io.hhplus.tdd.Integration.currency;
 
 import io.hhplus.tdd.database.PointHistoryTable;
 import io.hhplus.tdd.database.UserPointTable;
@@ -13,16 +13,16 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
 
-public class PointServiceSyncTest {
+public class PointServiceFOTest {
 
 
-    PointServiceSync pointService;
+    PointServiceFO pointService;
     @BeforeEach
     void setUp(){
         UserPointTable userPointTable = new UserPointTable();
         PointHistoryTable pointHistoryTable = new PointHistoryTable();
         ServiceValidation serviceValidation = new ServiceValidation();
-        pointService = new PointServiceSync(userPointTable,pointHistoryTable,serviceValidation);
+        pointService = new PointServiceFO(userPointTable,pointHistoryTable,serviceValidation);
     }
 
     private static final int THREAD_NUM = 100;
@@ -37,6 +37,7 @@ public class PointServiceSyncTest {
      * 같은 사용자가
      * 동시에 10_000 포인트 충전
      */
+    @DisplayName("같은 사용자가 한번에 스레드 수 만큼 충전할 때")
     @Test
     void sameUser_pointCharge() throws InterruptedException {
         ExecutorService executor = Executors.newFixedThreadPool(THREAD_NUM);
@@ -54,11 +55,11 @@ public class PointServiceSyncTest {
         long end_time = System.currentTimeMillis();
         timeRecorder(start_time,end_time);
 
-        Assertions.assertEquals(10_000*THREAD_NUM,pointService.findUserPointByUserId(1L).point());
+        Assertions.assertNotEquals(10_000*THREAD_NUM,pointService.findUserPointByUserId(1L).point());
         System.out.println("실제 저장된 포인트 : "+pointService.findUserPointByUserId(1L).point());
     }
 
-    @DisplayName("서로 다른 유저들이 각자 충전할때")
+    @DisplayName("쓰래드 수 만큼의 서로 다른 유저들이 각자 충전 할 때")
     /**
      * 전부 다른 유저 들로 할 경우
      * @throws InterruptedException
@@ -95,8 +96,5 @@ public class PointServiceSyncTest {
 
         Assertions.assertEquals(10_000*THREAD_NUM,totalPoint);
     }
-
-
-
 }
